Fix #334 and fix #308, add badge for onboarding
When building for Shield, this turns off the default opening of the sidebar on install. A badge is added to the toolbar button (with "New"), and on first click it will open the sidebar which will show onboarding. After clicking the button the badge will no longer be shown and the popup will appear as normal.